MUCH better system for packet pooling and multiplexing
This makes most things in `turbulence` abstract over the actual *packet* type
rather than just the buffer type.  This solves loads of problems.

First, it makes the mechanism for providing packet pools for the purposes of
buffer re-use a more easy to justify and explain, since it also allows the user
to just have any arbitrary packet type.

Much more importantly though, it allows the same trait for packet allocation to
be used for packet *multiplexing* by implementing a wrapper.  This means that
all of the code that previously only worked assuming `MuxPacket` now lo longer
depends on `MuxPacket` at all.  This will allow the user to pick and choose what
they want out of the library much more, even to the extent of only using
inner channel types directly.

Now that both of these things fall under the same trait system, the entire thing
is easier to justify and is much more worth it for the added complexity.