Don't 'nagle' at all, even in the reliable channels.
Also always reset the resend timer after doing a resend.

Firstly, it was *probably* technically wrong to call it nagle-ing because
nagle's algorithm I think also involves outstanding ACKs which this didn't do.

But in any case, it's confusing and not entirely helpful that *only* the
reliable channel will send unflushed data, just with a huge latency.  Since
flush is already required in the unreliable channel, this is simpler and should
be strictly better, because it avoids too frequent wakeups and avoids sending
too little data in the case that data is sent right before a flush would have
been triggered.