Document that all async methods are supposed to be cancel safe.
They are at least designed to be cancel safe, but it would be nice to
actually *test* this fact in all cases.

Not all of the async methods in the crate have added documentation, but at least
the ones where it wouldn't be *surprising* if they weren't cancel safe are
documented now.