Make unreliable channels more reliable, improve cancel guarantees.
The unreliable bincode channel would drop packets on send if they were
unable to be sent immediately, now instead use the same strategy as the
reliable bincode channel and record a pending write. This is better and
has virtually no additional overhead at all.

Also tighten the cancel safe guarantees of the bincode channels,
guaranteeing that if the `send` future is dropped before completion, no
message will be sent. This is an easy guarantee to make and is no worse
at all than not guaranteeing it, both types of bincode channel now just
don't call `finish_write` after queuing a sent message, relying on a
subsequent call to `send` or `flush` to call `finish_write`.