A lot of API and implementation simplification
* No more maximum message sizes for bincode channels, the maximum size
of messages is already u16::MAX and that is not a DoS risk at all.

* Since the conversion between bincode and non-bincode channels has no
extra settings, add the  ability to convert freely between them.

* Use `Vec` types as inner buffers in bincode channels because it is
almost exactly the same as using `Box<[u8]>` and an extra length, but
simpler and requires less space in some cases.

* `PacketPool` interface now reports a static maximum packet capacity
rather than each acquired packet. allowing things that use it to change
behavior based on a known static capacity.